requirements.txt
absl-py==0.10.0
astunparse==1.6.3
autocorrect==2.1.2
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
inexactsearch==1.0.2
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.1
matplotlib==3.3.2
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.3
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyspellchecker==0.5.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.11
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
silpa-common==0.3
six==1.15.0
sklearn==0.0
soundex==1.1.3
soupsieve==2.0.1
spellchecker==0.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.50.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
